% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coast_waste_10.R
\docType{data}
\name{coast_waste_10}
\alias{coast_waste_10}
\title{Annual Mismanaged Plastic Waste}
\format{
A data frame with 186 rows and 2 variables.
\describe{
\item{Entity}{Country or region name}
\item{Mismanaged plastic waste (tonnes)}{Tonnes of mismanaged plastic waste.}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2019/2019-05-21#coast_vs_waste.csv}
}
\usage{
coast_waste_10
}
\description{
This dataset provides an annual estimate of plastic discharge in the year 2010. It is worth noting that this
figure for each nation does not encompass waste exported abroad, which could have a higher likelihood of
ending up in the ocean.
}
\keyword{datasets}
