% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{descriptive_stats}
\alias{descriptive_stats}
\title{Summary_Statistics_Function}
\usage{
descriptive_stats(x)
}
\arguments{
\item{x}{: A numeric, character, or factor vector.}
}
\value{
A data frame with statistics such as frequency, proportion,
cumulative frequency, and other summary measures, depending on the input type.
}
\description{
This function summarizes either character/factor or numeric data, providing
frequency, proportion, cumulative frequency, and other statistics based on the
input type.
}
\examples{
library(tidyverse)
library(dvwaste)

# Example with a factor
factor_input <- factor(c("A", "B", "A", "C", "B"))
descriptive_stats(factor_input)

# Example with a character vector
character_input <- c("A", "B", "A", "C", "B")
descriptive_stats(character_input)

# Example with a numeric vector
numeric_input <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
descriptive_stats(numeric_input)

}
