% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvwaste-package.R
\docType{package}
\name{dvwaste-package}
\alias{dvwaste}
\alias{dvwaste-package}
\title{dvwaste: Global Plastic Waste Data Analysis and Visualisation}
\description{
Launch a Shiny App, load plastic pollution datasets and provide relevant functions to explore plastic pollution issues.
}
\author{
\strong{Maintainer}: Maliny Po \email{po0007@student.monash.com.au}

}
\keyword{internal}
