% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_plastic_waste.R
\docType{data}
\name{global_plastic_waste}
\alias{global_plastic_waste}
\title{Global Plastic Waste}
\format{
A data frame with 282 rows and 4 variables.
\describe{
\item{Entity}{Country or region name}
\item{Mismanaged_Waste}{Tonnes of mismanaged plastic waste.}
\item{Daily_Mismanaged_Waste}{Amount of mismanaged plastic waste per capita in kg/day.}
\item{Daily_Plastic_Waste}{Amount of plastic waste per capita in kg/day.}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2019/2019-05-21#coast_vs_waste.csv}
}
\usage{
global_plastic_waste
}
\description{
A dataset combining coast_waste_10 (Annual Mismanaged Plastic Waste), mismanaged_gdp_10 (Daily Per Capita Mismanaged Plastic Waste), and waste_gdp_10 (Daily Per Capita Plastic Waste).
}
\keyword{datasets}
