% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mismanaged_gdp_10.R
\docType{data}
\name{mismanaged_gdp_10}
\alias{mismanaged_gdp_10}
\title{Daily Per-Capita Mismanaged Plastic Waste}
\format{
A data frame with 282 rows and 2 variables.
\describe{
\item{Entity}{Country or region name}
\item{Per capita mismanaged plastic waste (kilograms per person per day)}{Amount of mismanaged plastic waste per capita in kg/day.}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2019/2019-05-21#coast_vs_wastecsv}
}
\usage{
mismanaged_gdp_10
}
\description{
For this dataset, mismanaged plastic waste refers to plastic that is either littered or improperly
disposed of. The total for each country does not include waste shipped overseas, where it may also
need to be appropriately managed.
}
\keyword{datasets}
