% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_barchart.R
\name{param_barchart}
\alias{param_barchart}
\title{Parameterised Bar Chart Function}
\usage{
param_barchart(country_list = character(0))
}
\arguments{
\item{country_list}{: A list/ vector of selected country.}
}
\value{
A table A tibble containing two columns: "Entity" (country names)
and "Value" (Mismanaged plastic waste in tonnes)
}
\description{
This function This feature filters and processes data from the 'coast_waste_10'
dataset for selected countries to generate a tibble suitable for constructing a bar chart.
}
\details{
Country element in the country_list must exist as an entry in Entity column from coast_waste_10 data set If left empty (default), data for all valid countries will be included.
}
\section{List of available \code{Entity} in coast_waste_10}{
available_countries <- coast_waste_10$Entity
available_countries
}

\examples{
library(tidyverse)
library(dvwaste)

# select multiple countries
input_countries <- c("Bahamas", "Australia", "United States")
param_barchart(input_countries)

# Use default country_list value, return a table with all countries
param_barchart()


}
