% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{run_app}
\alias{run_app}
\title{Launches a Shiny App from the 'dvwaste' package}
\usage{
run_app()
}
\value{
None (launches a Shiny App)
}
\description{
This function allows you to launch a Shiny web application that is bundled
with the 'dvwaste' package. It retrieves the application directory within
the package and runs it using Shiny's 'runApp' function.
}
\details{
The 'run_app' function is a convenient way to start the Shiny app provided
by the 'dvwaste' package. If the package is not installed or if the
application directory is not found, it will display an error message
suggesting a re-installation of the package.
}
\examples{
\dontrun{
# Run the Shiny app from 'dvwaste' package
run_app()
}

}
\seealso{
\code{\link[shiny]{runApp}}
}
